Move refreshcasestats to its own file
